+++ examples/angular/src/environments/environment.development.ts
@@ -5,6 +5,6 @@ export const environment = {
    * for security purposes, make sure is read-only.
    * More info: https://www.dotcms.com/docs/latest/permissions
    */
-  authToken: 'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhcGliMmYxMTYyYS01NDJmLTQwY2UtYjZmMi00NTEyOWFmODEwMmEiLCJ4bW9kIjoxNzE4MDM1MDIxMDAwLCJuYmYiOjE3MTgwMzUwMjEsImlzcyI6ImZlZjg2OTFlY2UiLCJsYWJlbCI6IkVNQSIsImV4cCI6MTgxMjYwMDAwMCwiaWF0IjoxNzE4MDM1MDIxLCJqdGkiOiIxYzcyN2QwNy0zNDM2LTQwMDAtYWZkOS00MmEwNjRmYTQ3ZWIifQ.bGet_1S5kpvWWNDtibBNasO6BsVPXvlq7wJaQQwQFiU',
+  authToken: 'YOUR_API_TOKEN',
   siteId: 'true',
 };
