+++ examples/nextjs/.env.local.example
@@ -1,4 +1,4 @@
-NEXT_PUBLIC_DOTCMS_AUTH_TOKEN=YOUR_READ_ONLY_API_TOKEN_HERE
+NEXT_PUBLIC_DOTCMS_AUTH_TOKEN=YOUR_API_TOKEN
 NEXT_PUBLIC_DOTCMS_HOST=http://localhost:8080
 NEXT_PUBLIC_EXPERIMENTS_API_KEY=analytic-api-key-from-dotcms-portlet
 NEXT_PUBLIC_EXPERIMENTS_DEBUG=true
